apps/api/pyproject.toml
[project]
name = "utah-research-matcher-api"
version = "0.1.0"
description = "FastAPI backend for the University of Utah company-to-researcher matcher."
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "beautifulsoup4>=4.13.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "openai>=1.99.0",
  "pgvector>=0.3.6",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.10.0",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.38",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.12.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
